refactor: legalRequirements.js を services/legal/ に分割（後方互換ファサード）
712行の単一ファイルを航空法/小型無人機法/土地管理者ルール別に分割し、
legalRequirements.js は re-export のみの薄いファサードに変更
